# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditOut": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Email"
          },
          "actor_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Role"
          },
          "actor_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor User Id"
          },
          "after": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "After"
          },
          "before": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "string"
          },
          "entity_type": {
            "title": "Entity Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "ip": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ip"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "id",
          "action",
          "entity_type",
          "entity_id",
          "actor_user_id",
          "actor_role",
          "before",
          "after",
          "request_id",
          "created_at"
        ],
        "title": "AuditOut",
        "type": "object"
      },
...
      "CaseDetailOut": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "closed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By"
          },
          "closed_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By Email"
          },
          "conflict_flag": {
            "title": "Conflict Flag",
            "type": "boolean"
          },
          "conflict_match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Match Result Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Email"
          },
          "duration_months": {
            "title": "Duration Months",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MatchOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Result Id"
          },
          "provider": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderBriefOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "provider_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Name"
          },
          "sanction_record": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SanctionRecordOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "sanction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanction Type"
          },
          "source_authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Authority"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subject_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Name"
          }
        },
        "required": [
          "id",
          "case_number",
          "provider_id",
          "sanction_record_id",
          "match_result_id",
          "status",
          "start_date",
          "end_date",
          "duration_months",
          "created_by",
          "closed_by",
          "close_reason",
          "conflict_flag",
          "conflict_match_result_id",
          "created_at",
          "provider_name",
          "subject_name",
          "sanction_type",
          "source_authority",
          "created_by_email",
          "closed_by_email",
          "provider",
          "sanction_record",
          "match"
        ],
        "title": "CaseDetailOut",
        "type": "object"
      },
      "CaseIn": {
        "additionalProperties": false,
        "properties": {
          "duration_months": {
            "anyOf": [
              {
                "maximum": 60.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to DEFAULT_CASE_MONTHS (3).",
            "title": "Duration Months"
          },
          "match_result_id": {
            "format": "uuid",
            "title": "Match Result Id",
            "type": "string"
          },
          "start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          }
        },
        "required": [
          "match_result_id"
        ],
        "title": "CaseIn",
        "type": "object"
      },
      "CaseListItemOut": {
        "additionalProperties": false,
        "description": "A case row with the names a person reads, not only the ids.",
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "closed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By"
          },
          "closed_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By Email"
          },
          "conflict_flag": {
            "title": "Conflict Flag",
            "type": "boolean"
          },
          "conflict_match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Match Result Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "created_by_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Email"
          },
          "duration_months": {
            "title": "Duration Months",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Result Id"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "provider_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Name"
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "sanction_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sanction Type"
          },
          "source_authority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Authority"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "subject_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject Name"
          }
        },
        "required": [
          "id",
          "case_number",
          "provider_id",
          "sanction_record_id",
          "match_result_id",
          "status",
          "start_date",
          "end_date",
          "duration_months",
          "created_by",
          "closed_by",
          "close_reason",
          "conflict_flag",
          "conflict_match_result_id",
          "created_at",
          "provider_name",
          "subject_name",
          "sanction_type",
          "source_authority",
          "created_by_email",
          "closed_by_email"
        ],
        "title": "CaseListItemOut",
        "type": "object"
      },
      "CaseOut": {
        "additionalProperties": false,
        "properties": {
          "case_number": {
            "title": "Case Number",
            "type": "string"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "closed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed By"
          },
          "conflict_flag": {
            "title": "Conflict Flag",
            "type": "boolean"
          },
          "conflict_match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Match Result Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "duration_months": {
            "title": "Duration Months",
            "type": "integer"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_result_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Result Id"
          },
          "provider_id": {
            "title": "Provider Id",
            "type": "string"
          },
          "sanction_record_id": {
            "format": "uuid",
            "title": "Sanction Record Id",
            "type": "string"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "case_number",
          "provider_id",
          "sanction_record_id",
          "match_result_id",
          "status",
          "start_date",
          "end_date",
          "duration_months",
          "created_by",
          "closed_by",
          "close_reason",
          "conflict_flag",
          "conflict_match_result_id",
          "created_at"
        ],
        "title": "CaseOut",
        "type": "object"
      },
      "CloseCaseIn": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "CloseCaseIn",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AuditOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[AuditOut]",
        "type": "object"
      },
      "Page_CaseListItemOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseListItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CaseListItemOut]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provider sanctions and exclusions reconciliation.\n\nMatches sanction records against a provider master file with a Fellegi-Sunter\nengine, EM-learned weights and isotonic calibration, and routes the grey band to\nan LLM adjudicator. Every run is pinned to a data snapshot and can be replayed.",
    "title": "Concordance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/cases": {
      "get": {
        "operationId": "list_cases_cases_get",
        "parameters": [
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "ACTIVE",
                    "EXPIRED",
                    "CLOSED",
                    "REJECTED"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "provider_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider Id"
            }
          },
          {
            "in": "query",
            "name": "conflict",
            "required": false,
            "schema": {
              "default": false,
              "title": "Conflict",
              "type": "boolean"
            }
          },
          {
            "description": "Start date on or after.",
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date on or after.",
              "title": "Date From"
            }
          },
          {
            "description": "Start date on or before.",
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Start date on or before.",
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CaseListItemOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Cases",
        "tags": [
          "cases"
        ]
      },
      "post": {
        "description": "Open a case on an approved match. Approval already opens one; this is for\nre-opening after that case closed or expired. Admin only.",
        "operationId": "create_cases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The match is not approved, or a case is already active."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}": {
      "get": {
        "description": "The case with its provider, the record it was opened on, and the approving match.",
        "operationId": "get_case_cases__case_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Case",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/audit": {
      "get": {
        "description": "Every recorded event on this case, newest first.",
        "operationId": "case_audit_cases__case_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "description": "Page size.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Page size.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Rows to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Rows to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Case Audit",
        "tags": [
          "cases"
        ]
      }
    },
    "/cases/{case_id}/close": {
      "post": {
        "description": "Close an active case, with a reason. Admin only.",
        "operationId": "close_cases__case_id__close_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Case Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseCaseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "The case is not active."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Close",
        "tags": [
          "cases"
        ]
      }
    },
...
  }
}